Add util.py for books/persions common function.
