+++ aozoracli/client.py
@@ -8,10 +8,16 @@ def main():
      print("Start aozora command line tool")
 
 def get_books(id=None, payload=None):
-    url = AOZORAPI_URL + "/books"
-    if id != None:
-        url += "/{}".format(id)
+    url = _create_api_url("books", id)
     return requests.get(url, params=payload)
 
-def get_persons(payload=None):
-    return requests.get(AOZORAPI_URL + "/persons", params=payload)
+def get_persons(id=None, payload=None):
+    url = _create_api_url("persons", id)
+    return requests.get(url, params=payload)
+
+def _create_api_url(path, id=None):
+    url = AOZORAPI_URL + "/" + path
+    if id != None:
+        url += "/{}".format(id)
+    return url
+
